refactor: split platform factory from platform classes
To be able to inherit from concrete platform implementations, we need to
make them importable. By that, we need to split-off the factory function
(which is only imported once) and rename the platform implementations so
that multiple can be imported in the same extension version.
